superagente/requirements.txt
langchain
langgraph
langchain_ollama
langchain_core
requests
pydantic
python-dotenv
